requirements.txt
streamlit
langchain
langchain_google_genai